backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
astroid==3.2.4
bcrypt==4.2.0
black==24.8.0
cffi==1.17.0
click==8.1.7
cryptography==43.0.0
dill==0.3.8
ecdsa==0.19.0
fastapi==0.112.0
greenlet==3.0.3
h11==0.14.0
idna==3.7
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.2.2
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pylint==3.2.6
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
ruff==0.5.6
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.37.2
tomlkit==0.13.0
typing_extensions==4.12.2
uvicorn==0.30.5
